Making a start on implementing rapidjson
Will be starting on implementing rapidjson into our project. For now I have created 2 text files, one which will be for all the decks in the game so we can access them all from one place and then a small placeholder deck list txt for the player which contains a few card objects in the correct format. Tomorrow I will be actually using rapidjson to read these files and create Card instances in Unity with the data gathered from them.