Began visuals for moving cards on playing field
Objects with the Card Controller attached can be moved around into Drop Zones.

For some reason, HandController allows the player to drop cards into it but PlayAreaController does not? I added a couple of funne kyle and josh pokemon cards as placeholders to see where the cards were going and how actual card art would look with that size.